Add files via upload
Based on the given data, we used independent component analysis, finding the interaction between different industry, i·e·, theres strong interaction between the boost of information industry and medical/healthcare industry, leading to a high stock return. This provide a scientific guidance for the policy makers for a greater outcome.